py_eda_nocv.py updated

+++ cdautil.py
@@ -44,9 +44,9 @@ def join_cmat(cmata, cmatb, ndimab):
     ndimb = cmatb.shape[0]
     noccb = cmatb.shape[1]
   
-    cmat_join[0:ndima/2,0:nocca] = cmata[0:ndima/2,0:nocca]
-    cmat_join[ndimab/2:ndimab/2+ndima/2,0:nocca] = cmata[ndima/2:,0:nocca]
-    cmat_join[ndima/2:ndimab/2,nocca:] = cmatb[0:ndimb/2,0:noccb]
-    cmat_join[ndimab/2+ndima/2: ,nocca:] = cmatb[ndimb/2:,0:noccb]
+    cmat_join[0:int(ndima/2),0:nocca] = cmata[0:int(ndima/2),0:nocca]
+    cmat_join[int(ndimab/2):int(ndimab/2+ndima/2),0:nocca] = cmata[int(ndima/2):,0:nocca]
+    cmat_join[int(ndima/2):int(ndimab/2),nocca:] = cmatb[0:int(ndimb/2),0:noccb]
+    cmat_join[int(ndimab/2+ndima/2): ,nocca:] = cmatb[int(ndimb/2):,0:noccb]
     
     return cmat_join
